Created a reset and solve button that will toggle. When 'Solve' is clicked, the board will display the solved puzzle and toggle 'Solve' to 'Reset?'. When 'Reset?' is clicked it will reset the board and toggle to 'Solve'. This works by running through the node list of allBoardTiles in solveBoard() and filling in each tile using the index of that tile (found in the id) and filling it with the boardSolution array index.
To reset the game will run through the nodelist of board tiles (tiles in resetGame()) and will remove each one. This is done in a while loop that won't stop until there are no more elements in the nodelist. Then we style the buttons accordingly to toggle back to 'Solve'.